--- a/tests/ert/unit_tests/analysis/test_es_update.py
+++ b/tests/ert/unit_tests/analysis/test_es_update.py
@@ -403,7 +403,7 @@ def test_smoother_snapshot_alpha(
                     "index": polars.Series(range(len(data)), dtype=polars.UInt16),
                     "values": data,
                 }
-            ),
+            ).pivot(on="response_key", index=["report_step", "index"]),
             iens,
         )
     posterior_storage = storage.create_ensemble(
@@ -736,7 +736,7 @@ def test_gen_data_obs_data_mismatch(storage, uniform_parameter):
                     "index": polars.Series(range(len(data)), dtype=polars.UInt16),
                     "values": polars.Series(data, dtype=polars.Float32),
                 }
-            ),
+            ).pivot(on="response_key", index=["report_step", "index"]),
             iens,
         )
     posterior_ens = storage.create_ensemble(
@@ -799,7 +799,7 @@ def test_gen_data_missing(storage, uniform_parameter, obs):
                     "index": polars.Series(range(len(data)), dtype=polars.UInt16),
                     "values": polars.Series(data, dtype=polars.Float32),
                 }
-            ),
+            ).pivot(on="response_key", index=["report_step", "index"]),
             iens,
         )
     posterior_ens = storage.create_ensemble(
@@ -893,7 +893,7 @@ def test_update_subset_parameters(storage, uniform_parameter, obs):
                     "index": polars.Series(range(len(data)), dtype=polars.UInt16),
                     "values": polars.Series(data, dtype=polars.Float32),
                 }
-            ),
+            ).pivot(on="response_key", index=["report_step", "index"]),
             iens,
         )
     posterior_ens = storage.create_ensemble(