--- a/tests/ert/unit_tests/storage/test_storage_migration.py
+++ b/tests/ert/unit_tests/storage/test_storage_migration.py
@@ -178,9 +178,9 @@ def test_that_storage_matches(
             tuple(ensemble.get_realization_list_with_responses()),
         )
         snapshot.assert_match(
-            summary_data.sort("time", "response_key", "realization")
+            summary_data.sort("time", "realization")
             .to_pandas()
-            .set_index(["time", "response_key", "realization"])
+            .set_index(["time", "realization"])
             .transform(np.sort)
             .to_csv(),
             "summary_data",
@@ -197,9 +197,9 @@ def test_that_storage_matches(
             "gen_data", tuple(range(ensemble.ensemble_size))
         )
         snapshot.assert_match(
-            gen_data.sort(["realization", "response_key", "report_step", "index"])
+            gen_data.sort(["realization", "report_step", "index"])
             .to_pandas()
-            .set_index(["realization", "response_key", "report_step", "index"])
+            .set_index(["realization", "report_step", "index"])
             .to_csv(),
             "gen_data",
         )