… tests/ert/unit_tests/storage/snapshots/test_storage_migration/test_that_storage_matches/gen_data
… tests/ert/unit_tests/storage/snapshots/test_storage_migration/test_that_storage_matches/summary_data